+++ app/src/main/java/org/breezyweather/sources/mf/json/MfForecastDaily.kt
@@ -24,8 +24,8 @@ import java.util.Date
 @Serializable
 data class MfForecastDaily(
     @Serializable(DateSerializer::class) val time: Date,
-    @SerialName("T_min") val tMin: Float?,
-    @SerialName("T_max") val tMax: Float?,
+    @SerialName("T_min") val tMin: Double?,
+    @SerialName("T_max") val tMax: Double?,
     @SerialName("daily_weather_icon") val dailyWeatherIcon: String?,
     @SerialName("daily_weather_description") val dailyWeatherDescription: String?,
     @SerialName("sunrise_time") @Serializable(DateSerializer::class) val sunriseTime: Date?,
