+++ app/src/main/java/org/breezyweather/sources/openmeteo/json/OpenMeteoWeatherCurrent.kt
@@ -21,17 +21,17 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class OpenMeteoWeatherCurrent(
-    @SerialName("temperature_2m") val temperature: Float?,
-    @SerialName("apparent_temperature") val apparentTemperature: Float?,
+    @SerialName("temperature_2m") val temperature: Double?,
+    @SerialName("apparent_temperature") val apparentTemperature: Double?,
     @SerialName("weathercode") val weatherCode: Int?,
-    @SerialName("windspeed_10m") val windSpeed: Float?,
-    @SerialName("winddirection_10m") val windDirection: Float?,
-    @SerialName("windgusts_10m") val windGusts: Float?,
-    @SerialName("uv_index") val uvIndex: Float?,
+    @SerialName("windspeed_10m") val windSpeed: Double?,
+    @SerialName("winddirection_10m") val windDirection: Double?,
+    @SerialName("windgusts_10m") val windGusts: Double?,
+    @SerialName("uv_index") val uvIndex: Double?,
     @SerialName("relativehumidity_2m") val relativeHumidity: Int?,
-    @SerialName("dewpoint_2m") val dewPoint: Float?,
-    @SerialName("pressure_msl") val pressureMsl: Float?,
+    @SerialName("dewpoint_2m") val dewPoint: Double?,
+    @SerialName("pressure_msl") val pressureMsl: Double?,
     @SerialName("cloudcover") val cloudCover: Int?,
-    val visibility: Float?,
+    val visibility: Double?,
     val time: Long
 )
