+++ app/src/main/java/org/breezyweather/sources/openmeteo/json/OpenMeteoWeatherDaily.kt
@@ -22,11 +22,11 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class OpenMeteoWeatherDaily(
     val time: LongArray,
-    @SerialName("temperature_2m_max") val temperatureMax: Array<Float?>?,
-    @SerialName("temperature_2m_min") val temperatureMin: Array<Float?>?,
-    @SerialName("apparent_temperature_max") val apparentTemperatureMax: Array<Float?>?,
-    @SerialName("apparent_temperature_min") val apparentTemperatureMin: Array<Float?>?,
+    @SerialName("temperature_2m_max") val temperatureMax: Array<Double?>?,
+    @SerialName("temperature_2m_min") val temperatureMin: Array<Double?>?,
+    @SerialName("apparent_temperature_max") val apparentTemperatureMax: Array<Double?>?,
+    @SerialName("apparent_temperature_min") val apparentTemperatureMin: Array<Double?>?,
     val sunrise: Array<Long?>?,
     val sunset: Array<Long?>?,
-    @SerialName("uv_index_max") val uvIndexMax: Array<Float?>?
+    @SerialName("uv_index_max") val uvIndexMax: Array<Double?>?
 )
