+++ app/src/main/java/org/breezyweather/sources/openmeteo/json/OpenMeteoWeatherMinutely.kt
@@ -22,5 +22,5 @@ import kotlinx.serialization.Serializable
 data class OpenMeteoWeatherMinutely(
     val time: LongArray,
     //@SerialName("precipitation_probability") val precipitationProbability: Array<Int?>?,
-    val precipitation: Array<Float?>?
+    val precipitation: Array<Double?>?
 )
