+++ app/src/main/java/org/breezyweather/sources/openweather/json/OpenWeatherOneCallDailyTemp.kt
@@ -20,10 +20,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class OpenWeatherOneCallDailyTemp(
-    val day: Float?,
-    val min: Float?,
-    val max: Float?,
-    val night: Float?,
-    val eve: Float?,
-    val morn: Float?
+    val day: Double?,
+    val min: Double?,
+    val max: Double?,
+    val night: Double?,
+    val eve: Double?,
+    val morn: Double?
 )
