+++ app/src/main/java/org/breezyweather/sources/smhi/SmhiApi.kt
@@ -24,7 +24,7 @@ import retrofit2.http.Path
 interface SmhiApi {
     @GET("category/pmp3g/version/2/geotype/point/lon/{lon}/lat/{lat}/data.json")
     fun getForecast(
-        @Path("lon") lon: Float,
-        @Path("lat") lat: Float
+        @Path("lon") lon: Double,
+        @Path("lat") lat: Double
     ): Observable<SmhiForecastResult>
 }
